Add files via upload
KNN
The K-Nearest Neighbors (KNN) algorithm is a simple and straightforward method used in machine learning for classification and regression tasks.

KNN works by looking at the nearest neighbors of something new and deciding what it should be based on what most of those neighbors are. It’s like asking your friends for advice on something new by seeing what they think.
Imagine you have a bunch of objects that are already sorted into groups. For example, let's say you have a bunch of fruits, and each fruit is labeled as either an apple, an orange, or a banana.
Now, you find a new fruit, and you want to figure out what it is. You can use the KNN algorithm to help you. Here’s how:
Look Around: You check out the closest fruits to this new one. Let's say you decide to look at the 3 closest fruits (this number is called "K" in KNN).
Count the Neighbors: You see that 2 of these nearby fruits are apples, and 1 is a banana.
Make a Decision: Since most of the nearby fruits are apples, you decide that your new fruit is also likely to be an apple.